Implementation for issue #50: https://github.com/amethyst/evoli/issues/50
Details:
= replaced time_control.rs with main_game_ui.rs
+ added main_game.ron, a prefab for the main game ui
+ added 'Menu' button to main_game_ui.rs
+ handle 'Menu' button in main_game.rs: transition to MenuState
= synchronized "ui style" between menu.ron and main_game.ron (i.e. button colors)
+ included macros from 'log' crate in main.rs; annotated some program flow with log's info! macro
+ added MainGameState::on_stop() to clean up world upon returning to the MenuState

Known Bugs:
After returning to the Menu and clicking Start, the program panics at line 72 of decision.rs trying to unwrap a faction that resolves as None. I *do* delete all entities in MainGameState::on_stop(), but I don't believe my code is the cause of this panic, just that it exposes the flaw.