Add carcasses on death events (#103)
Add a death event that triggers on starvation and death by
health loss. If the dying entity has a new carcass component, spawn
an entity of the creature type held in that component. Remove
nutrtition to fullness transfer on attack death, and instead decrease
nutrition of the attacked by the attack damage of the attacker. This
allows for entities that do not spawn carcasses (i.e. plants) to still
provide nutrition to their attacker, and prevents live creatures
(i.e. herbivores) that do spawn carcasses from providing any nutrition
while alive.