Fix loading Wind resource from config file
load() now returns a Result, so updated the LoadingState accordingly. Should the resource fail to load, this will be logged in the console and the default() method will be used instead.

Also changed the Default implementation for Wind; it used to be (0, 0) or perfectly windless, which is a poor default as it does not adequately demo the Topplegrass entities.